refactor: (WIP, bridge) singleton pattern
